Make POSTing KeyAuthorization optional, V2 don't echo it.
This commit updates the RA to make the notion of submitting
a `KeyAuthorization` value as part of the `ra.UpdateAuthorization` call
optional. If set, the value is enforced against expected and an error is
returned if the provided authorization isn't correct. If it isn't set
the RA populates the field with the computed authorization for the VA to
enforce against the value it sees in challenges.

The ACMEv2/WFEv2 `prepChallengeForDisplay` function is updated to strip
the `ProvidedKeyAuthorization` field before sending the challenge object
back to a client. ACMEv1/WFEv1 continue to return the KeyAuthorization
in challenges to avoid breaking clients that are relying on this legacy
behaviour.

For deployability ease this commit retains the name of the
`core.Challenge.ProvidedKeyAuthorization` field even though it _should_
be called `core.Challenge.ComputedKeyAuthorization` now that it isn't
set based on the client's provided key authz. This will be easier as
a follow-up change.