ACMEv2: Enforce POST Content-Type (when feature on)
This commit adds a new WFE2 feature flag "EnforceV2ContentType". When
enabled, the WFE2's `validPostRequest` function will enforce that the
request carries a `Content-Type` header equal to
`application/jose+json`. This is required by ACME draft-10 per section
6.2 "Request Authentication".

This is behind a feature flag because it is likely to break
some number of existing ACMEv2 clients that may not be sending the
correct `Content-Type`.