Harmonize order expiry with assoc. authz expiry.
Prior to this commit an order's expiry was set based on
`ra.orderLiftime` while pending and valid authorization expiry was set
based on `ra.pendingAuthorizationLifetime` and
`ra.authorizationLifetime`. Since orders reused existing valid/pending
authorizations this can lead to a case where an order has an expiry
beyond the associated authorization expiries. In this case when an
authorization expired the order becomes inactionable and the extra order
lifetime is not useful.

This commit addresses this problem in two ways:
1) The SA `GetAuthorizations` function used to find authzs to reuse for
   `ra.NewOrder` is adjusted to only return authorizations at min 24hr
   away from expiry.
2) Order expiry is now calculated as the min of the order's own expiry
   or the soonest authorization expiry. This properly reflects the
   order's true lifetime based on the authorization lifetime.

The RA/SA unit tests are updated accordingly.