RA: Include authz IDs and challenge types in `certificateRequestEvent`.
This commit updates the RA such that `certificateRequestEvent` objects
created during issuance also include
a `map[string]certificateRequestAuthz` map that can be used to map from
a certificate name to an authorization ID and the challenge type that
made the authorization valid. A unit test that checks that the expected
authorization IDs and challenge types are set is included.