CI: Fix 1.10 tag, review feedback.
Prior to this commit the `tag_and_upload.sh` script was using `1.10` to
refer to `1.10.0`. Upstream `1.10` was aliased to `1.10.1`. We need to
be explicit and use `1.10.0` to get `1.10.0`. This commit fixes this
oversight (and additionally updates the `Dockerfile.tmpl`, which lost
the magic `%%GO_VERSION%%` accidentally).

This commit removes the `-e GO_VERSION` from the docker-compose
invocation because it isn't actually needed in the containers.

This commit also removes the use of `sudo` in the `.travis.yml`'s
`before_install` to install an updated `docker-compose` and instead puts
the binary in a non-root location and runs it via absolute path.