RA: Don't lose CA error types when prefixing err msg.
Previously we updated the RA's `issueCertificateInner` function to
prefix errors returned from the CA with meaningful information about
which CA RPC caused the failure. Unfortunately by using `fmt.Errorf` to
do this we're discarding the underlying error type. This can cause
unexpected server internal errors downstream if (for e.g.) the CA
rejects a CSR with a malformed error.

This commit updates the `issueCertificateInner` error message prefixing
to maintain the error type if the underlying error is
a `berrors.BoulderError`. A RA unit test with several mock CAs is added
to test the prefixing occurs as expected without loss of error type.

This commit also adds an integration test that ensures we reject a CSR
with >100 names with a malformed error. To allow this test to pass the
`pendingAuthorizationsPerAccount` rate limit in
`test/rate-limit-policies.yml` has to be bumped (and the unit tests that
match this content updated).