gRPC: publish RPC latency stat in server interceptor.
We may see RPCs that are dispatched by a client but do not arrive at the
server for some time afterwards. To have insight into potential request
latency at this layer we want to publish the time delta between when
a client sent an RPC and when the server received it.

This commit updates the gRPC client interceptor to add the current time
to the gRPC request metadata context when it dispatches an RPC. The
server side interceptor is updated to pull the client request time out
of the gRPC request metadata. Using this timestamp it can calculate
the latency and publish it as an observation on a Prometheus histogram.

Accomplishing the above required wiring a clock through to each of the
client interceptors. This caused a small diff across each of the gRPC
aware boulder commands.

A small unit test is included in this commit that checks that a latency
stat is published to the histogram after an RPC to a test ChillerServer
is made.