load generator: send correct ACMEv2 Content-Type on POST.
This commit updates the Boulder `load-generator` to send the correct
ACMEv2 `Content-Type` header when POSTing the ACME server. This is
required for ACMEv2 and without it all POST requests to the WFE2 running
a `test/config-next` configuration result in malformed 400 errors. While
only required by ACMEv2 this commit sends it for ACMEv1 requests as
well. No harm no foul.