integration tests: verify load-generator req success.
Prior to this commit we ran the load generator in integration tests
against both the V1 and the V2 API for a short duration. This helps
prevent bitrot in the load-generator from catching us unaware. At the
time of landing this code we also tried to verify that none of the
load-generator requests failed. This was found to be flaky/unreliable at
the time and removed.

This commit re-adds the check that all of the load-generator requests
were successful. We've made some nice performance/reliability
improvements lately and I think we can now expect that the
load-generator will produce no errors. This will help catch regressions
in the load-generator that do not result in a non-zero exit code but
still indicate a problem that should be addressed.