gRPC: publish in-flight RPC gauge in client interceptor.
This commit updates the Boulder gRPC `clientInterceptor` to publish
a Prometheus gauge stat for each in-flight RPC it dispatches, sliced by
service and method.

A unit test is included that uses a custom `ChillerServer` that lets the
test block up a bunch of RPCs, check the in-flight gauge value is
increased, unblock the RPCs, and recheck that the in-flight gauge is
reduced. To check the gauge value for a specific set of labels a new
`test-tools.go` function `GaugeValueWithLabels` is added.