SA: Cleanup, forbid nil issuer arg to AddCertificate
In https://github.com/letsencrypt/boulder/pull/3651 we introduced a new
parameter to `sa.AddCertificate` to allow specifying the Issued date. If
`nil`, we defaulted to the current time to maintain deployability
guidelines.

Now that this has been deployed everywhere this commit updates
`SA.AddCertificate` and the gRPC wrappers such that a nil `issuer`
argument is rejected with an error.

Unit tests that were previously using `nil` for the issued time are
updated to explicitly set the issued time to the fake clock's `now()`.