WFE2: Allow legacy Key ID prefix for ACME v2 JWS.
While we intended to allow legacy ACME v1 accounts created through the
WFE to work with the ACME v2 implementation and the WFE2 we neglected to
account for the fact that a legacy account would have a Key ID URL that
doesn't match the expected for a V2 account. This caused
`wfe2/verify.go`'s `lookupJWK` to reject all POST requests authenticated
by a legacy account unless the ACME client took the extra manual step of
"fixing" the URL.

This commit adds a configuration parameter to the WFE2 for an allowed
legacy key ID. The WFE2 verification logic is updated to allow both the
expected key ID and the configured legacy key ID. This will allow us to
specify the correct legacy URL in configuration for both staging/prod to
allow unmodified V1 ACME accounts to be used with ACME v2.