Fix issue with expired authz test.
The test_expired_authz_404() test is currently broken in two ways - firstly,
there is no way for it to distinguish between a 404 from an expired authz
and a 404 from a non-existant authz. Secondly, the test_expired_authz_purger()
test runs and wipes out all of the existing authorizations, including the one
that was set up from setup_seventy_days_ago(), before the expired test runs.

Avoid this by running the tests in order by name and rename the purger test
to force it to be run last. Also, add a minimal canary that will detect if
all authorizations have been purged (although this still does not guarantee
that we got a 404 due to expiration).