VA: Log challenge type in `checkCAA`.
This commit updates the VA's `checkCAA` function to include the provided
`challengeType` parameter as part of the audit log line for the CAA
check result. If `challengeType == nil` then the value "none" is logged.
Unit tests are updated accordingly.

This change will make it easier to distinguish cases where "Valid for
issuance" is false because of a validation-methods restriction.