VA: Log exceptional non-problem remote VA errors.
Prior to this commit if the primary VA encountered a non-problem error
return from a remote VA as a result of a `PerformValidation` RPC during
`performRemoteValidation` a server internal problem without the
underlying error in the detail was returned as the result. This masks
the true error and makes debugging from logs difficult.

This commit updates the primary VA `performRemoteValidation` function to
log non-nil non-problem error results from a call to a remote VA's
`PerformValidation` function at the error log level.

A unit test with a mock remote VA that always returns an unexpected
non-problem error from `PerformValidation` is used to test that the
expected user-facing server internal problem is returned from
`performRemoteValidation` and that an audit error log with the true
underlying error is written.