README: Remove slowstart instructions.
None of the development team use this approach to running a dev env. It
no longer works without modifying `test/startservers.py` and the `test/`
configurations. Given that it has been broken for a month+ without
and has only provoked one user issue I think we can be fairly confident
that few others are using this method of setting up a Boulder
development environment.