SA: statusForOrder shouldn't fetch authzs for expired orders.
If an order is expired the status is invalid and we don't need to get
any of the order's authorizations. Its important to exit early in this
case because an order that references an expired authorization will be
itself have been expired (because we match the order expiry to the
associated authz expiries in `ra.NewOrder`), and expired authorizations
may be purged from the DB. Because of this purging fetching the authz's
for an expired order may return less authz objects than expected,
triggering a 500 error response.