Filter invalid UTF-8 from error responses.
For HTTP-01 challenges that return incorrect responses, the
VA tries to put the first little bit of the HTTP response in the problem
detail.

However, VA needs to be able to serialize the problem detail as a
protobuf to send it to the RA, and protobufs require string types to be
UTF-8. Since providing some response body is only a convenience anyhow,
when the response body is not valid UTF-8, we simply don't make it part
of the error detail.