Docs: Update/Restructure REAMDE and assoc. docs
This is primarily based on Roland's `readme-updates` branch and the
closed PR[0].

The README is restructured to be user-driven. Information that is
strictly contributor focused (e.g. dep upgrades) is moved to
CONTRIBUTING.md.

The README now links to the production deployment guide in the wiki and
describes our general position on using Boulder.

The CONTRIBUTING.md guide links to the various docs/ pages and was
updated.

The docs/acme-divergences page is updated for draft-15.

The DESIGN.md document is moved from the root of the repo into docs/.
It has been updated to cover ACMEv2 and precertificates/SCT embedding.

[0] - https://github.com/letsencrypt/boulder/pull/3195