setup cleanup: update build.sh, remove setup.sh (#3885)
`test/setup.sh` is removed because it was a relic of the "slowstart" instructions we removed from the README previously. One less place to keep in-sync.

`test/boulder-tools/build.sh` is updated to fix three problems: 
  1) the `github.com/golang/lint/golint` package moved to `golang.org/x/lint/golint`
  2) it did not abort if a background process (e.g. `apt` or `go get`) failed for any reason.
  3) Certbot master has removed tls-alpn-01 support

Problem 1 was causing `go get github.com/golang/lint/golint` to fail. Problem 2 was causing Problem 1 to be ignored, resulting in Docker images that were missing all of the dev tools (missing `goose` is the
first error to indicate this state).

Problem 3 breaks the TLS-ALPN-01 integration tests from `test/config-next` and requires using a specific version of Certbot (even though we don't rely on Certbot for solving the TLS-ALPN-01 challenge, we do rely on it to POST the initiation of the challenge and that does not function with Certbot master).